requirements.txt
# Core HTTP & feeds
requests>=2.31.0
feedparser>=6.0.10

# Geospatial
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0
fiona>=1.9.0

# AI APIs
mistralai>=1.0.0
elevenlabs>=1.0.0

# Environment & config
python-dotenv>=1.0.0

# UI
gradio>=4.0.0

# Scheduling
apscheduler>=3.10.0

# Testing
pytest>=7.4.0
pytest-mock>=3.11.0
responses>=0.23.0

# Notebooks
jupyter>=1.0.0
matplotlib>=3.7.0
folium>=0.14.0
